[MISC][RL] Decouple animation from variant
