+++ src/shared/internal-calendar/internal-calendar.tsx
@@ -1,7 +1,5 @@
 import dayjs, { Dayjs } from "dayjs";
 import React, { useEffect, useImperativeHandle, useRef, useState } from "react";
-import { useResizeDetector } from "react-resize-detector";
-import { useSpring } from "react-spring";
 import { CalendarHelper } from "../../util/calendar-helper";
 import { InternalCalendarDay } from "./internal-calendar-day";
 import { InternalCalendarMonth } from "./internal-calendar-month";
@@ -9,7 +7,6 @@ import { InternalCalendarYear } from "./internal-calendar-year";
 import {
     ActionButton,
     ActionButtonSection,
-    AnimatedDiv,
     ArrowLeft,
     ArrowRight,
     Container,
@@ -42,7 +39,6 @@ export const Component = (
         onHover,
         onDismiss,
         actionComponent,
-        isOpen,
         value,
         endValue,
         currentFocus,
@@ -66,7 +62,7 @@ export const Component = (
 
     const doneButtonRef = useRef<HTMLButtonElement>(null);
     const cancelButtonRef = useRef<HTMLButtonElement>(null);
-    const resizeDetector = useResizeDetector();
+    const containerRef = useRef<HTMLDivElement>();
 
     // =============================================================================
     // HOOKS
@@ -87,17 +83,18 @@ export const Component = (
     // =============================================================================
     // EFFECTS
     // =============================================================================
-    useEffect(() => {
-        // open with 'confirmed' value for day calendar in first mounted
-        if (!isOpen) return;
+    // TODO:
+    // useEffect(() => {
+    //     // open with 'confirmed' value for day calendar in first mounted
+    //     if (!isOpen) return;
 
-        setCalendarDate(dayjs());
-        setCurrentView("default");
+    //     setCalendarDate(dayjs());
+    //     setCurrentView("default");
 
-        if (variant === "range" && value?.length && endValue?.length) {
-            setIsNewSelection(false);
-        }
-    }, [isOpen]);
+    //     if (variant === "range" && value?.length && endValue?.length) {
+    //         setIsNewSelection(false);
+    //     }
+    // }, [isOpen]);
 
     useEffect(() => {
         /**
@@ -280,7 +277,7 @@ export const Component = (
             performOnCalendarView("month-options");
 
             // Maintain focus when selecting month dropdown
-            resizeDetector.ref.current.focus();
+            containerRef.current.focus();
         } else {
             setCurrentView("default");
             setCalendarDate(viewCalendarDate);
@@ -467,69 +464,54 @@ export const Component = (
         );
     };
 
-    const renderContent = () => {
-        return (
-            <Container
-                ref={resizeDetector.ref}
-                tabIndex={-1}
-                data-id="calendar-container"
-                $type={type}
-                {...otherProps}
-            >
-                {type === "standalone" && (
-                    <SideArrowButton
-                        $direction="left"
-                        onClick={handleLeftArrowClick}
-                    >
-                        <ArrowLeft />
-                    </SideArrowButton>
-                )}
-                <ContentBody>
-                    {renderHeader()}
-                    <ToggleZone>
-                        <InternalCalendarDay
-                            type={type}
-                            calendarDate={calendarDate}
-                            currentFocus={currentFocus}
-                            disabledDates={disabledDates}
-                            selectedStartDate={selectedStartDate}
-                            selectedEndDate={selectedEndDate}
-                            variant={variant}
-                            between={between}
-                            isNewSelection={isNewSelection}
-                            onSelect={handleDateSelect}
-                            onHover={handleHover}
-                        />
-                        <OptionsOverlay $visible={currentView !== "default"}>
-                            {renderOptionsOverlay()}
-                        </OptionsOverlay>
-                    </ToggleZone>
-                    {renderActionButtons()}
-                </ContentBody>
-                {type === "standalone" && (
-                    <SideArrowButton
-                        $direction="right"
-                        onClick={handleRightArrowClick}
-                    >
-                        <ArrowRight />
-                    </SideArrowButton>
-                )}
-            </Container>
-        );
-    };
-
-    // React spring animation configuration
-    const styles = useSpring({
-        height: isOpen
-            ? resizeDetector.height + 64 // include vertical padding
-            : 0,
-    });
-
-    if (type === "input") {
-        return <AnimatedDiv style={styles}>{renderContent()}</AnimatedDiv>;
-    } else {
-        return <>{renderContent()}</>;
-    }
+    return (
+        <Container
+            ref={containerRef}
+            tabIndex={-1}
+            data-id="calendar-container"
+            $type={type}
+            {...otherProps}
+        >
+            {type === "standalone" && (
+                <SideArrowButton
+                    $direction="left"
+                    onClick={handleLeftArrowClick}
+                >
+                    <ArrowLeft />
+                </SideArrowButton>
+            )}
+            <ContentBody>
+                {renderHeader()}
+                <ToggleZone>
+                    <InternalCalendarDay
+                        type={type}
+                        calendarDate={calendarDate}
+                        currentFocus={currentFocus}
+                        disabledDates={disabledDates}
+                        selectedStartDate={selectedStartDate}
+                        selectedEndDate={selectedEndDate}
+                        variant={variant}
+                        between={between}
+                        isNewSelection={isNewSelection}
+                        onSelect={handleDateSelect}
+                        onHover={handleHover}
+                    />
+                    <OptionsOverlay $visible={currentView !== "default"}>
+                        {renderOptionsOverlay()}
+                    </OptionsOverlay>
+                </ToggleZone>
+                {renderActionButtons()}
+            </ContentBody>
+            {type === "standalone" && (
+                <SideArrowButton
+                    $direction="right"
+                    onClick={handleRightArrowClick}
+                >
+                    <ArrowRight />
+                </SideArrowButton>
+            )}
+        </Container>
+    );
 };
 
 export const InternalCalendar = React.forwardRef(Component);
